backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1
numpy>=1.24.0
google-generativeai==0.3.2
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.0
aiohttp==3.9.1
httpx
